Rearranged some files
Moved delete_stack to scripts folder
deleted duplicate create_stack in cloudformation folder
